Extract Beat Card selection functionality into separate file as extension of HomeScreenViewController
